Fix FontAwesome inclusion
Joomla components are made to use a custom FA distribution
which replaces the fa- prefix with icon-. This is
distributed in the default Joomla templates, Cassiopeia and
Atum.

Lightning was including the FontAwesome proper CSS, also
distributed with Joomla but NOT used by the core components
because any kind of consistency would be un-Joomla-like.

So here I am, using the custom FontAwesome CSS from the
Cassiopeia template to avoid weird display issues. Sigh